// client/vscode/src/webview/platform/context.ts
export interface VSCodePlatformContext
    extends Pick<
        PlatformContext,
        | 'updateSettings'
        | 'settings'
        | 'getGraphQLClient'
        | 'getStaticExtensions'
        | 'telemetryService'
        | 'telemetryRecorder'
        | 'clientApplication'
    > {
    // Ensure telemetryService is non-nullable.
    telemetryService: VsceTelemetryService
    requestGraphQL: <R, V = object>(options: {
        request: string
        variables: V
        mightContainPrivateInfo: boolean
        overrideAccessToken?: string
        overrideKhulnasoftURL?: string
    }) => Observable<GraphQLResult<R>>
}